Moved `Variable` to top level
